refactor(threat-detection): extract provider result mappers
