+++ packages/core/src/threatDetection/lakera/index.ts
@@ -9,6 +9,7 @@
  */
 
 export { LakeraClient, type LakeraClientConfig } from "./LakeraClient";
+export { toLakeraThreatDetectionResult } from "./toThreatDetectionResult";
 export {
   lakeraGuardRequestSchema,
   lakeraGuardResponseSchema,
